# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Chat Database Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MongoConnectionInput {
  /** MongoDB connection string (mongodb:// or mongodb+srv://) */
  uri: string;
}

export interface MongoConnectionResult {
  ok: boolean;
  dbName: string;
  host: string;
}

export interface CollectionInfo {
  name: string;
  type: string;
  /** @nullable */
  count?: number | null;
}

export interface CollectionListResult {
  collections: CollectionInfo[];
}

export interface CreateCollectionInput {
  uri: string;
  /** Name of the collection to create */
  name: string;
}

export interface CreateCollectionResult {
  ok: boolean;
  name: string;
  message?: string;
}

export type CollectionSetupStatusStatus = typeof CollectionSetupStatusStatus[keyof typeof CollectionSetupStatusStatus];


export const CollectionSetupStatusStatus = {
  created: 'created',
  already_exists: 'already_exists',
  error: 'error',
} as const;

export interface CollectionSetupStatus {
  name: string;
  status: CollectionSetupStatusStatus;
  /** @nullable */
  error?: string | null;
}

export interface SetupResult {
  collections: CollectionSetupStatus[];
  created: number;
  skipped: number;
  errors: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListCollectionsParams = {
uri: string;
};