# lib/api-zod/src/generated/types/collectionSetupStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Chat Database Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionSetupStatusStatus } from './collectionSetupStatusStatus';

export interface CollectionSetupStatus {
  name: string;
  status: CollectionSetupStatusStatus;
  /** @nullable */
  error?: string | null;
}