# lib/api-zod/src/generated/types/setupResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Chat Database Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionSetupStatus } from './collectionSetupStatus';

export interface SetupResult {
  collections: CollectionSetupStatus[];
  created: number;
  skipped: number;
  errors: number;
}